fix linter errors

+++ contracts/diamond/upgradeInitializers/DiamondMultiInit.sol
@@ -18,10 +18,10 @@ contract DiamondMultiInit {
     // The `diamondCut` function executes this function to execute multiple initializer functions for a single upgrade.
 
     function multiInit(address[] calldata _addresses, bytes[] calldata _calldata) external {
-        if(_addresses.length != _calldata.length) {
+        if (_addresses.length != _calldata.length) {
             revert AddressAndCalldataLengthDoNotMatch(_addresses.length, _calldata.length);
         }
-        for(uint i; i < _addresses.length; i++) {
+        for (uint256 i; i < _addresses.length; i++) {
             LibDiamond.initializeDiamondCut(_addresses[i], _calldata[i]);
         }
     }
